Split annotation resolution to discord.utils
